Monorepo, refactors for better login and chunking.
In /:
	Setup monorepo config with NPM workspaces for easier project
management. Server package is for the backend, web is for the frontend,
and types is for lib stuff like shared types and helper functions.

In server/:
	Did many refactors. Setup login sessions with cookies. Setup
better handlebars template rendering for output web bundle.

In web/:
	Setup vite chunking for better site load times. Setup SASS
BootStrap usage without importing all of BootStrap for every file by
using SASS partials and importing to the main SASS file instead of the
React componets.

In types/:
	Added more helper functions, and improved some typings.