Fixed Dockerfile and symlink refactors.
In Dockerfile:
	Debugged docker image and make multi-stage build for all major
testing steps of any needed images. Now has development and production
builds. The production build is optimised to use less space, and it
stores the main backend files in server/ instead of /dist-tsc.

In server/package.json:
	Removed some unused packages.

In server/*:
	Fixed dotenv to use a symlink to aviod failing to find .env file
depending on the current direcctory Node.js was started from. And added
some debuging output for the dotenv .env file parsing.

In .dockerignore and .gitignore:
	Added some more excluded files, and fixed some patterns to be
more specific.